+++ src/main/scala/onextent/bluecql/gen/DbCode.scala
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package onextent.bluecql
+package onextent.bluecql.gen
 
 import org.apache.cassandra.cql3.statements.CreateTableStatement
